refactor(cli): split tests.rs into per-domain test modules
The cli/ split landed in d509680 left tests.rs at 1466 LOC, violating
the no-source-file-over-1000-LOC rule the PR was supposed to enforce.

Split into cli/tests/{exit_code,args_parsing,cache_ops,warm_lockfile,
daemon,util,snapshot,analyze}.rs grouped by the cli/ submodule each
test targets. Each new file is well under 1K LOC (largest: analyze.rs
at 494). Helpers (make_test_store, write_lockfile, make_journal_line*,
write_fixture_journal, default_opts) live alongside the tests that
use them — no shared helper module needed.

All 65 unit tests still pass; no behavior change.
